test-requirements.txt
aws-cdk-lib==2.20.0
pytest~=7.1.1
pytest-cov>=3.0.0
